Content Atom - ensure js/main.js is up to date.
The js/main.js file is missing quite a lot of the content atoms that should be there while the js library is languishing quite far behind at 2.1.1 to the 2.4.7 we're currently on. I've updated the version to 2.4.7 to bring in line rather than bumping from what it was previously on. Please shout if you'd prefer I did not do this.