Deprecate suppressFurntiture field froms Flags.
The suppressFurniture field was added to the model as part of an intended feature but it never got implemented. I've double checked and we have no atoms in Elasticsearch with this field. This also breaks our concept of having fields on our data which make little sense and relate to a UI.