Add Subtitles asset type
We considered that in future, we might also want to model a Captions asset type. We debated whether a Text asset type would be more generic, but we would then have to add an optional field on the Asset to handle whether it was Captions or Subtitles (which would be awkward as it wouldn't apply to videos and audio).

We also discussed how we could model the language of the subtitles. We could model this directly on the Asset, rather than AssetType. A Language field could just as easily apply to Video and Audio as it would to Subtitles. This is not needed for now - but you never know!

For these reasons we think going for a Subtitles AssetType is a suitable approach.

Co-authored-by: Andy Noton <andy.noton.contractor@guardian.co.uk>